+++ src/main/java/net/telepathicgrunt/bumblezone/features/CaveSugarWaterfall.java
@@ -11,6 +11,7 @@ import net.minecraft.world.gen.chunk.ChunkGenerator;
 import net.minecraft.world.gen.feature.DefaultFeatureConfig;
 import net.minecraft.world.gen.feature.Feature;
 import net.telepathicgrunt.bumblezone.blocks.BzBlocks;
+import net.telepathicgrunt.bumblezone.tags.BZBlockTags;
 
 import java.util.Random;
 
@@ -29,7 +30,7 @@ public class CaveSugarWaterfall extends Feature<DefaultFeatureConfig> {
         BlockPos.Mutable blockpos$Mutable = new BlockPos.Mutable().set(position);
         BlockState blockstate = world.getBlockState(blockpos$Mutable.up());
 
-        if (!blockstate.isOpaque() || Registry.BLOCK.getId(blockstate.getBlock()).getPath().contains("honey")) {
+        if (!blockstate.isOpaque() || blockstate.isIn(BZBlockTags.HONEYCOMBS_THAT_FEATURES_CAN_CARVE)) {
             return false;
         } else {
             //checks if we are in the side of a wall with air exposed on one side
@@ -38,7 +39,7 @@ public class CaveSugarWaterfall extends Feature<DefaultFeatureConfig> {
             int neededNumberOfSides;
             blockstate = world.getBlockState(blockpos$Mutable.down());
 
-            if (blockstate.isOpaque() && Registry.BLOCK.getId(blockstate.getBlock()).getPath().contains("honey")) {
+            if (blockstate.isOpaque() && blockstate.isIn(BZBlockTags.HONEYCOMBS_THAT_FEATURES_CAN_CARVE)) {
                 neededNumberOfSides = 3;
             } else if (blockstate.getBlock() == CAVE_AIR.getBlock()) {
                 neededNumberOfSides = 4;
@@ -49,7 +50,7 @@ public class CaveSugarWaterfall extends Feature<DefaultFeatureConfig> {
 
             for (Direction face : Direction.Type.HORIZONTAL) {
                 blockstate = world.getBlockState(blockpos$Mutable.offset(face));
-                if (blockstate.isOpaque() && Registry.BLOCK.getId(blockstate.getBlock()).getPath().contains("honey")) {
+                if (blockstate.isOpaque() && blockstate.isIn(BZBlockTags.HONEYCOMBS_THAT_FEATURES_CAN_CARVE)) {
                     ++numberOfSolidSides;
                 } else if (blockstate.getBlock() != CAVE_AIR.getBlock()) {
                     return false;