+++ src/main/java/net/telepathicgrunt/bumblezone/tags/BZBlockTags.java
@@ -11,4 +11,5 @@ public class BZBlockTags {
 
     public static final Tag.Identified<Block> REQUIRED_BLOCKS_UNDER_HIVE_TO_TELEPORT_TAG = BlockTagsAccessor.bz_callRegister(Bumblezone.MODID + ":required_blocks_under_hive_to_teleport");
     public static final Tag.Identified<Block> BLACKLISTED_TELEPORTATION_HIVES_TAG = BlockTagsAccessor.bz_callRegister(Bumblezone.MODID+":blacklisted_teleportable_hive_blocks");
+    public static final Tag.Identified<Block> HONEYCOMBS_THAT_FEATURES_CAN_CARVE = BlockTagsAccessor.bz_callRegister(Bumblezone.MODID+":honeycombs_that_features_can_carve");
 }